Callback funtion changed to accept long integer as length of data

+++ gcc-plugin/uPrint.h
@@ -1,3 +1,3 @@
 #include <stdarg.h>
 
-extern void uPrint(void (*cb)(int, unsigned char *), char *fmt, ...);
+extern void uPrint(void (*cb)(unsigned long, unsigned char *), char *fmt, ...);
